requirements.txt
# Minimal deps to run tests/demo (no Rust builds)
python-dotenv>=1.1.0
flask>=3.0.0
openai>=1.0.0

# Testing dependencies
pytest>=8.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality dependencies
pylint>=3.0.0
pytest-pylint>=0.21.0
coverage>=7.0.0
